test: gate the MCP wire-capture test to unix platforms
The test launches a shell-script agent, which cannot run on
Windows: the Windows CI job failed with exec: "/bin/sh":
executable file not found. Move it into session_unix_test.go
with a //go:build unix constraint, matching the same package's
session_teardown_test.go and the sibling agent packages' script
tests. The cmd/sortie coverage scan parses raw source, so its
callsite requirement still holds on every platform.
