+++ internal/agent/clientprotocol/session_unix_test.go
@@ -1,0 +1,78 @@
+//go:build unix
+
+package clientprotocol
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/sortie-ai/sortie/internal/agent/agenttest"
+	"github.com/sortie-ai/sortie/internal/domain"
+	"github.com/sortie-ai/sortie/internal/registry"
+)
+
+// mcpHandshakeScript is a fake agent that answers exactly the two
+// calls startSession makes before returning: initialize (always id 1,
+// since it is the connection's first call) and session/new (always id
+// 2). It captures the raw session/new request line to captureFile
+// before answering it, so a test can inspect the exact bytes the
+// adapter wrote for its own tool-server delivery.
+func mcpHandshakeScript(captureFile string) string {
+	return `capture='` + captureFile + `'
+while IFS= read -r line; do
+  case "$line" in
+    *'"method":"initialize"'*)
+      printf '%s\n' '{"jsonrpc":"2.0","id":1,"result":{"protocolVersion":1,"agentCapabilities":{"mcpCapabilities":{"http":true}}}}'
+      ;;
+    *'"method":"session/new"'*)
+      printf '%s\n' "$line" >> "$capture"
+      printf '%s\n' '{"jsonrpc":"2.0","id":2,"result":{"sessionId":"sess-1"}}'
+      ;;
+  esac
+done
+`
+}
+
+// TestStartSessionMCPInjectionWire calls agenttest.AssertMCPInjection
+// with the wire channel populated from the actual session/new request
+// bytes the adapter wrote, and is what gives cmd/sortie's
+// TestEveryAgentKindHasMCPInjectionCoverage a call site for this
+// kind. It launches a shell-script agent, so it runs only where
+// /bin/sh exists.
+func TestStartSessionMCPInjectionWire(t *testing.T) {
+	t.Parallel()
+
+	dir := t.TempDir()
+	mcpConfigPath := writeMCPConfig(t)
+	capturePath := filepath.Join(dir, "session_new.jsonl")
+	scriptPath := agenttest.WriteScript(t, dir, "agent.sh", mcpHandshakeScript(capturePath))
+
+	params := domain.StartSessionParams{
+		WorkspacePath: t.TempDir(),
+		AgentConfig:   domain.AgentConfig{Command: scriptPath},
+		MCPConfigPath: mcpConfigPath,
+	}
+
+	session, err := startSession(context.Background(), params)
+	if err != nil {
+		t.Fatalf("startSession() error = %v", err)
+	}
+	t.Cleanup(func() { _ = stopSession(context.Background(), session) })
+
+	captured, err := os.ReadFile(capturePath)
+	if err != nil {
+		t.Fatalf("read captured session/new request: %v", err)
+	}
+
+	declared, ok := registry.Agents.Meta("agent-client-protocol")
+	if !ok {
+		t.Fatal(`registry.Agents.Meta("agent-client-protocol") ok = false, want true`)
+	}
+
+	agenttest.AssertMCPInjection(t, declared.MCPInjection, mcpConfigPath, agenttest.MCPLaunchSurface{
+		Wire: []string{strings.TrimSpace(string(captured))},
+	})
+}
